apps: greybus-utils: fix build breakage
Fix build breakage due to last minute change from size_t to u16, which
is not a type that is available. Apparently failed to compile test for
the right target before committing.

Note that the endianess macros that have been used in the svc
implementation also rely on u16 when CONFIG_ENDIAN_BIG is enabled...

Testing Done: tested on BDB1B
Bug Number: trivial
Review URL: https://reviews.leaflabs.com/r/762/
Reviewed by:

Reported-by: Darryl Newberry <newberry_darryl@projectara.com>

+++ apps/greybus-utils/svc_msg.c
@@ -61,7 +61,7 @@ void send_hot_plug(char *hpe, int iid)
     struct svc_msg *msg = (struct svc_msg *)hpe;
     struct greybus_manifest_header *mh =
         (struct greybus_manifest_header *)(hpe + HP_BASE_SIZE);
-    u16 manifest_size = le16toh(mh->size);
+    size_t manifest_size = le16toh(mh->size);
 
     msg->header.function_id = SVC_FUNCTION_HOTPLUG;
     msg->header.message_type = SVC_MSG_DATA;
